libap: fix 32-bit time overflow throughout (2038 + long-sleep correctness)
On Plan9 6c, long=32-bit even on amd64.  Any time_t/tv_sec value * 1000
overflows after ~24 days; any seconds-since-epoch stored as long or int
wraps in 2038.  Fixes from a full audit of all libap time arithmetic:

sys/time.h:
  struct timeval.tv_sec: long → long long (matches time_t; tv_usec stays
  long since microseconds always fit in 32 bits)

time/nanosleep.c:
  int ms → long long ms; explicit (long long) cast before *1000

plan9/__p9_syscall.c:
  p9_nanosleep: long ms → long long ms; (long long) cast before *1000
  gettimeofday: remove (long) narrowing cast on tv_sec — tv_sec is now
  long long so the cast was silently truncating post-2038 timestamps

select/poll.c (ppoll):
  int timeout → long long timeout_ms intermediate; clamp to INT_MAX
  before passing to poll() (poll takes int ms; kernel limit is ~24 days)

plan9/_buf.c (select):
  int t → long long tms intermediate; clamp to INT_MAX before storing
  in Muxbuf.waittime (int) and passing to _SLEEP (long) — both 32-bit
  kernel interfaces

thread/cond_timedwait.c:
  long ns → long long ns; explicit (long long) casts in multiplication;
  critical because any wait > ~2 seconds previously overflowed silently

stat/utimens.c:
  Add comment: Dir.atime/mtime (ulong) are a 9P wire protocol limitation;
  the (ulong) cast is intentional — not fixable without a kernel change

+++ sys/src/ape/lib/ap/plan9/__p9_syscall.c
@@ -369,10 +369,10 @@ p9_getuid(void)
 static long
 p9_nanosleep(const struct timespec *req, struct timespec *rem)
 {
-	long ms;
+	long long ms;
 	if(req == NULL)
 		return -EFAULT;
-	ms = (long)(req->tv_sec * 1000 + req->tv_nsec / 1000000);
+	ms = (long long)req->tv_sec * 1000 + req->tv_nsec / 1000000;
 	if(ms < 0) ms = 0;
 	if(_SLEEP(ms) < 0){
 		if(rem != NULL){
@@ -697,7 +697,7 @@ __p9_syscall(long n, long a1, long a2, long a3, long a4, long a5, long a6)
 		long long ns;
 		if(tv == NULL) return -EFAULT;
 		ns = nsec();
-		tv->tv_sec  = (long)(ns / 1000000000LL);
+		tv->tv_sec  = ns / 1000000000LL;
 		tv->tv_usec = (long)((ns % 1000000000LL) / 1000LL);
 		return 0;
 	}
