+++ sys/src/ape/lib/ap/time/nanosleep.c
@@ -7,9 +7,9 @@
 int
 nanosleep(struct timespec *req, struct timespec *rem)
 {
-	int ms;
+	long long ms;
 
-	ms = req->tv_sec * 1000 + (req->tv_nsec + 999999) / 1000000;
+	ms = (long long)req->tv_sec * 1000 + (req->tv_nsec + 999999) / 1000000;
 	if(_SLEEP(ms) < 0) {
 		if(rem) {
 			rem->tv_sec = rem->tv_nsec = 0; /* needs better handling */
